refactor(kernel): nest rpc_transport into core/ per §4
Mirror of the cas_* move (`dc9ad261d`). `rpc_transport.rs` is the
kernel's out-bound VFS gRPC client, used by `core::meta_store::
remote::RemoteMetaStore`, `core::pipe::remote::RemotePipeBackend`,
and `core::stream::remote::RemoteStreamBackend` — every internal
consumer already lives in `core/`. The file itself sat at the
kernel crate root.

Move `rust/kernel/src/rpc_transport.rs` → `core/rpc_transport.rs`,
declare `pub mod rpc_transport` in `core/mod.rs`, and replace the
crate-root `pub mod rpc_transport;` with `pub use core::rpc_transport;`
so `kernel::rpc_transport::RpcTransport` keeps resolving for the
peer crates (`transport`, `backends::storage::remote`).

Same seven-criteria pass as the CAS move: simpler contract (one
home for primitives), architecturally consistent with §4, SSOT
preserved, Rust-first, zero runtime cost, systematic.

`cargo check --workspace --tests` clean.
