+++ rust/kernel/src/lib.rs
@@ -95,10 +95,9 @@ pub mod abi;
 
 // Client-side RPC transport for `RemoteBackend` (the
 // `backends::storage::remote::RemoteBackend` ObjectStore impl that
-// proxies all syscalls over gRPC to a remote `nexusd`). The driver-
-// layer `rpc` crate re-exports this module as `rpc::vfs` so peer
-// crates name a single canonical path; the file lives here in the
-// kernel because the kernel-internal `RemoteMetaStore` /
-// `RemotePipeBackend` / `RemoteStreamBackend` wrappers also wrap
-// `RpcTransport` directly.
-pub mod rpc_transport;
+// proxies all syscalls over gRPC to a remote `nexusd`). Implementation
+// lives in `core::rpc_transport` per §4 alongside the kernel-internal
+// `RemoteMetaStore` / `Remote{Pipe,Stream}Backend` wrappers that wrap
+// it; re-exported here under the historical flat name so peer crates
+// (`transport`, `backends`) keep their existing import paths.
+pub use core::rpc_transport;
